Support servers being started on-demand by launchd
In on-demand mode, launchd reads the .plist file and listens on the
specified Unix domain sockets / TCP ports / other things. When something
interesting happens, launchd will spawn the service and expect it to
call `launch_activate_sockets` to receive the listening sockets. The
client is expected to call `accept` on these sockets.

This is very similar to the existing socket server, except that launchd
is the one who creates and listens on the sockets.

Signed-off-by: David Scott <dave.scott@unikernel.com>