Favor resolving into IPv4 over IPv6
If the server is listening to both it would make little difference.  If the
server is listening only to IPv4, as does cohttp, then it makes the endpoint
usable. If the server is listening only to IPv6, as does no HTTP server I know
about, then the endpoint is now broken.

A better fix would be to keep all addresses and try them all in turn, as
suggested in #244.