lwt-unix: refine the closing logic to support half-shutdowns
The input and output channels share the same underlying fd, and
so close(2) cannot be called twice on them (or else we run the risk
of closing a recycled fd or receiving an EBADF exception).

We therefore track the half-close state of the channels, and call
shutdown(2) on the underlying fd.  close(2) is only called if the
other half of the channel pair has been shutdown.