conduit-lwt-unix: fail when hostname verification cannot be done on SSL
When a valid hostname is not available it's better to fail early with
a useful error message rather than letting the connection go on and
letting OpenSSL fail with an undecipherable message.

Note that the "hostname" parameters are strings and don't have to be
hostnames, they can be IPs as well when using cohttp. Ideally these
should be a union type of domain names and ip addresses for better
clarity, but this would be a breaking change.

Signed-off-by: Pau Ruiz Safont <pau.safont@citrix.com>